+++ Implementation/src/calculator.c
@@ -11,7 +11,7 @@ int substract(int number1, int number2)
     return number1-number2;
 }
 
-int multiply(int number1, int number2)
+double multiply(double number1, double number2)
 {
     return number1*number2;
 }
